+++ jbpm-wb-integration/jbpm-wb-integration-client/src/main/java/org/jbpm/workbench/wi/client/editors/deployment/descriptor/sections/configuration/DeploymentsConfigurationView.java
@@ -27,7 +27,7 @@ import elemental2.dom.HTMLTableSectionElement;
 import org.jboss.errai.ui.shared.api.annotations.DataField;
 import org.jboss.errai.ui.shared.api.annotations.EventHandler;
 import org.jboss.errai.ui.shared.api.annotations.Templated;
-import org.kie.workbench.common.screens.library.client.settings.sections.SectionView;
+import org.kie.workbench.common.screens.library.client.settings.util.sections.SectionView;
 
 @Templated
 public class DeploymentsConfigurationView implements SectionView<DeploymentsConfigurationPresenter> {
